+++ src/django_routines/__init__.py
@@ -1,16 +1,23 @@
 r"""
 ::
 
-        ____    _                            ____              __  _
-       / __ \  (_)___ _____  ____ _____     / __ \____  __  __/ /_(_)___  ___  _____
-      / / / / / / __ `/ __ \/ __ `/ __ \   / /_/ / __ \/ / / / __/ / __ \/ _ \/ ___/
-     / /_/ / / / /_/ / / / / /_/ / /_/ /  / _, _/ /_/ / /_/ / /_/ / / / /  __(__  )
-    /_____/_/ /\__,_/_/ /_/\__, /\____/  /_/ |_|\____/\__,_/\__/_/_/ /_/\___/____/
-         /___/            /____/
+        ██████╗      ██╗ █████╗ ███╗   ██╗ ██████╗  ██████╗
+        ██╔══██╗     ██║██╔══██╗████╗  ██║██╔════╝ ██╔═══██╗
+        ██║  ██║     ██║███████║██╔██╗ ██║██║  ███╗██║   ██║
+        ██║  ██║██   ██║██╔══██║██║╚██╗██║██║   ██║██║   ██║
+        ██████╔╝╚█████╔╝██║  ██║██║ ╚████║╚██████╔╝╚██████╔╝
+        ╚═════╝  ╚════╝ ╚═╝  ╚═╝╚═╝  ╚═══╝ ╚═════╝  ╚═════╝
+
+    ██████╗  ██████╗ ██╗   ██╗████████╗██╗███╗   ██╗███████╗███████╗
+    ██╔══██╗██╔═══██╗██║   ██║╚══██╔══╝██║████╗  ██║██╔════╝██╔════╝
+    ██████╔╝██║   ██║██║   ██║   ██║   ██║██╔██╗ ██║█████╗  ███████╗
+    ██╔══██╗██║   ██║██║   ██║   ██║   ██║██║╚██╗██║██╔══╝  ╚════██║
+    ██║  ██║╚██████╔╝╚██████╔╝   ██║   ██║██║ ╚████║███████╗███████║
+    ╚═╝  ╚═╝ ╚═════╝  ╚═════╝    ╚═╝   ╚═╝╚═╝  ╚═══╝╚══════╝╚══════╝
 
 
-A simple Django app that allows you to specify batches of commands in your settings files
-and then run them in sequence by name using the provied ``routine`` command.
+A simple Django app that allows you to specify batches of commands in your settings
+files and then run them in sequence by name using the provided ``routine`` command.
 """
 
 import bisect
@@ -23,7 +30,7 @@ from dataclasses import asdict, dataclass, field
 from django.core.exceptions import ImproperlyConfigured
 from django.utils.functional import Promise
 
-VERSION = (1, 4, 0)
+VERSION = (1, 5, 0)
 
 __title__ = "Django Routines"
 __version__ = ".".join(str(i) for i in VERSION)
@@ -40,6 +47,9 @@ __all__ = [
     "command",
     "get_routine",
     "routines",
+    "Command",
+    "PreHook",
+    "PostHook",
 ]
 
 
@@ -49,6 +59,44 @@ ROUTINE_SETTING = "DJANGO_ROUTINES"
 R = t.TypeVar("R")
 CommandTypes = t.Union[t.Type["ManagementCommand"], t.Type["SystemCommand"]]
 Command = t.Union["ManagementCommand", "SystemCommand"]
+"""
+Command type, either a ManagementCommand or SystemCommand.
+"""
+
+_Hook = t.Callable[
+    [
+        "Routine",
+        "Command",
+        t.Optional["Command"],
+        t.Dict[str, t.Any],
+    ],
+    t.Optional[bool],
+]
+
+PreHook = _Hook
+"""
+Function type signature for a pre-hook functions. Pre-hook functions can modify command
+objects (including their arguments) before they are run.
+
+:param routine: The routine being run.
+:param command: The command about to be run.
+:param previous: The previous command that was run (or None if at the beginning).
+:param options: A dictionary containing the routine level options.
+:return: Return true to skip the command.
+"""
+
+PostHook = _Hook
+"""
+Function type signature for a post-hook functions. Post-hook functions can modify
+command objects (including their results) after they are run or the next command
+before it is run. Returning a truthy value will exit the routine early.
+
+:param routine: The routine being run.
+:param command: The command about to be run.
+:param next: The next command that will be run (or None if at the end).
+:param options: A dictionary containing the routine level options.
+:return: Return true to exit the routine early.
+"""
 
 
 def to_symbol(name: str, check_keyword: bool = False) -> str:
@@ -86,19 +134,27 @@ class _RoutineCommand(ABC):
     or for all invocations of the routine if no switches are configured.
     """
 
-    # todo
-    # pre_hook: t.Optional[t.Callable[[Command], t.Optional[bool]]] = None
-    # """
-    # A function to run before the command is run. The function should take the command instance and
-    # may return True to skip the command.
-    # """
+    pre_hook: t.Optional[PreHook] = None
+    """
+    A function that will be run before the command is run. It may make modifications
+    to the command or decide to skip the command by returning True. See :attr:`PreHook`
+    """
 
-    # post_hook: t.Optional[t.Callable[[Command, t.Any], t.Any]] = None
-    # """
-    # A function to run after the command has been run. The function should take the command instance
-    # and the result of the function which will be whatever call_command returns if run in the same
-    # process, or the result object returned by subprocess.run() if run in a subprocess.
-    # """
+    post_hook: t.Optional[PostHook] = None
+    """
+    A function that will be run after the command has been run. It may make
+    modifications to the command (including its result) or decide to exit the routine
+    early by returning True. See :attr:`PostHook`
+
+    It may also make modifications to the next command that will be run, if any.
+    """
+
+    result: t.Any = None
+    """
+    The result of the command run. This will either be the value returned by
+    :func:`~django.core.management.call_command` or a :func:`subprocess.run` result
+    object if the command was run in a subprocess.
+    """
 
     def __post_init__(self):
         self.switches = tuple([to_symbol(switch) for switch in self.switches])
@@ -143,7 +199,8 @@ class _RoutineCommand(ABC):
 class ManagementCommand(_RoutineCommand):
     options: t.Dict[str, t.Any] = field(default_factory=dict)
     """
-    t.Any options to pass to the command via call_command. Not valid for SystemCommands
+    Any options to pass to the command via
+    :func:`~django.core.management.call_command`. Not valid for SystemCommands
     """
 
     @property
@@ -170,8 +227,8 @@ def _insort_right_with_key(a: t.List[R], x: R, key: t.Callable[[R], t.Any]) -> N
     """
     A function that implements bisect.insort_right with a key callable on items.
 
-    todo: remove this and replace with key argument to bisect.insort_right when support for
-    python <3.10 dropped.
+    todo: remove this and replace with key argument to bisect.insort_right when support
+    for python <3.10 dropped.
     """
     transformed_list = [key(item) for item in a]
     transformed_x = key(x)
@@ -217,6 +274,21 @@ class Routine:
     Keep going if a command fails.
     """
 
+    pre_hook: t.Optional[PreHook] = None
+    """
+    This function will be run before each command that lacks its own pre_hook in the
+    routine. See :attr:`PreHook`. You can determine if this is the starting command
+    of the routine by checking if the previous command is None.
+    """
+
+    post_hook: t.Optional[PostHook] = None
+    """
+    This function will be run after each command that lacks its own post_hook in the
+    routine. See :attr:`PostHook`. You can determine if this is the last command of the
+    routine by checking if the next command is None. Post hooks may also be used to
+    exit the routine early by returning True.
+    """
+
     def __post_init__(self):
         self.name = to_symbol(self.name)
         self.switch_helps = {
@@ -235,8 +307,19 @@ class Routine:
         return sorted(switches)
 
     def plan(self, switches: t.Set[str]) -> t.List[Command]:
+        def set_hooks(
+            command: Command,
+            pre_hook: t.Optional[PreHook],
+            post_hook: t.Optional[PostHook],
+        ) -> Command:
+            if pre_hook and not command.pre_hook:
+                command.pre_hook = pre_hook
+            if post_hook and not command.post_hook:
+                command.post_hook = post_hook
+            return command
+
         return [
-            command
+            set_hooks(command, self.pre_hook, self.post_hook)
             for command in self.commands
             if not command.switches
             or any(switch in switches for switch in command.switches)
@@ -278,6 +361,8 @@ class Routine:
             "subprocess": self.subprocess,
             "atomic": self.atomic,
             "continue_on_error": self.continue_on_error,
+            "pre_hook": self.pre_hook,
+            "post_hook": self.post_hook,
         }
 
 
@@ -288,6 +373,8 @@ def routine(
     subprocess: bool = False,
     atomic: bool = False,
     continue_on_error: bool = False,
+    pre_hook: t.Optional[PreHook] = None,
+    post_hook: t.Optional[PostHook] = None,
     **switch_helps,
 ):
     """
@@ -296,6 +383,15 @@ def routine(
     :param name: The name of the routine to register.
     :param help_text: The help text to display for the routine by the routines command.
     :param commands: The commands to run in the routine.
+    :param subprocess: If true run each of the commands in a subprocess.
+    :param atomic: Run all commands in the same transaction.
+    :param continue_on_error: Keep going if a command fails.
+    :param pre_hook: A function to run before each command in the routine is run if the
+        command does not have its own pre_hook. See :attr:`PreHook`
+    :param post_hook: A function to run after each command in the routine is run if the
+        command does not have its own post_hook. See :attr:`PostHook`
+    :param switch_helps: A mapping of switch names to help text for the switches.
+    :raises: ImproperlyConfigured if the DJANGO_ROUTINES settings variable is not valid.
     """
     settings = sys._getframe(1).f_globals
     if not settings.get(ROUTINE_SETTING, {}):
@@ -322,6 +418,8 @@ def routine(
         subprocess=subprocess,
         atomic=atomic,
         continue_on_error=continue_on_error,
+        pre_hook=pre_hook,
+        post_hook=post_hook,
     )
 
     for command in commands:
@@ -351,6 +449,8 @@ def _add_command(
     *command: str,
     priority: int = _RoutineCommand.priority,
     switches: t.Optional[t.Sequence[str]] = _RoutineCommand.switches,
+    pre_hook: t.Optional[PreHook] = None,
+    post_hook: t.Optional[PostHook] = None,
     **options,
 ):
     settings = sys._getframe(2).f_globals
@@ -363,7 +463,12 @@ def _add_command(
     extra = {"options": options} if command_type is ManagementCommand else {}
     new_cmd = routine_obj.add(
         command_type(
-            t.cast(t.Tuple[str], command), priority, tuple(switches or []), **extra
+            t.cast(t.Tuple[str], command),
+            priority,
+            tuple(switches or []),
+            pre_hook=pre_hook,
+            post_hook=post_hook,
+            **extra,
         )
     )
     routine_dict[routine] = routine_obj.to_dict()
@@ -375,6 +480,8 @@ def command(
     *command: str,
     priority: int = RoutineCommand.priority,
     switches: t.Optional[t.Sequence[str]] = RoutineCommand.switches,
+    pre_hook: t.Optional[PreHook] = None,
+    post_hook: t.Optional[PostHook] = None,
     **options,
 ):
     """
@@ -391,7 +498,12 @@ def command(
         insertion order (defaults to zero).
     :param switches: The command will run only when one of these switches is active,
         or for all invocations of the routine if no switches are configured.
-    :param options: t.Any options to pass to the command via call_command.
+    :param options: t.Any options to pass to the command via
+        :func:`~django.core.management.call_command`.
+    :param pre_hook: A function to run before the command is run. See :attr:`PreHook`
+    :param post_hook: A function to run after the command has been run. See
+        :attr:`PostHook`
+    :raises: ImproperlyConfigured if the DJANGO_ROUTINES settings variable is not valid.
     :return: The new command.
     """
     return _add_command(
@@ -400,6 +512,8 @@ def command(
         *command,
         priority=priority,
         switches=switches,
+        pre_hook=pre_hook,
+        post_hook=post_hook,
         **options,
     )
 
@@ -409,6 +523,8 @@ def system(
     *command: str,
     priority: int = _RoutineCommand.priority,
     switches: t.Optional[t.Sequence[str]] = _RoutineCommand.switches,
+    pre_hook: t.Optional[PreHook] = None,
+    post_hook: t.Optional[PostHook] = None,
 ):
     """
     Add a system command to the named routine in settings to be run.
@@ -420,14 +536,24 @@ def system(
     :param routine: The name of the routine the command belongs to.
     :param command: The command and its arguments to run the routine, all strings or
         coercible to strings that the command will parse correctly.
-    :param priority: The order of the command in the routine. Priority ties will be run in
-        insertion order (defaults to zero).
+    :param priority: The order of the command in the routine. Priority ties will be run
+        in insertion order (defaults to zero).
     :param switches: The command will run only when one of these switches is active,
         or for all invocations of the routine if no switches are configured.
+    :param pre_hook: A function to run before the command is run. See :attr:`PreHook`
+    :param post_hook: A function to run after the command has been run. See
+        :attr:`PostHook`
+    :raises: ImproperlyConfigured if the DJANGO_ROUTINES settings variable is not valid.
     :return: The new command.
     """
     return _add_command(
-        SystemCommand, routine, *command, priority=priority, switches=switches
+        SystemCommand,
+        routine,
+        *command,
+        priority=priority,
+        switches=switches,
+        pre_hook=pre_hook,
+        post_hook=post_hook,
     )
 
 
@@ -437,7 +563,8 @@ def get_routine(name: str) -> Routine:
 
     .. note::
 
-        If called before settings have been configured, this function must be called from settings.
+        If called before settings have been configured, this function must be called
+        from settings.
 
     :param name: The name of the routine to get.
     :return: The routine.