+++ tests/django_routines_tests/management/commands/rtrack.py
@@ -1,0 +1,7 @@
+from .track import Command as TrackCommand
+
+
+class Command(TrackCommand):
+    def handle(self, *args, **options) -> str:
+        super().handle(*args, **options)
+        return str(options["id"])