+++ tests/system_cmd.py
@@ -9,7 +9,11 @@ import json
 
 if __name__ == "__main__":
     if not track_file.is_file():
-        track_file.write_text(json.dumps({"invoked": [], "passed_options": []}))
+        track_file.write_text(
+            json.dumps({"invoked": [], "passed_options": []}, indent=4)
+        )
     track = json.loads(track_file.read_text())
     track["invoked"].append(sys.argv[-1])
-    track_file.write_text(json.dumps(track))
+    track_file.write_text(json.dumps(track, indent=4))
+    if len(sys.argv) > 1:
+        print(sys.argv[-1])